+++ ui/src/reusable_ui/index.ts
@@ -1,6 +1,6 @@
 // Import Components
 import Button from './components/Button'
-import Dropdown from './components/dropdowns/Dropdown'
+import Dropdown, {DropdownMode} from './components/dropdowns/Dropdown'
 import MultiSelectDropdown from './components/dropdowns/MultiSelectDropdown'
 import Form from './components/form_layout/Form'
 import Input, {InputType} from './components/inputs/Input'
@@ -30,6 +30,7 @@ import {
 export {
   Button,
   Dropdown,
+  DropdownMode,
   MultiSelectDropdown,
   Form,
   Input,